campusai-rag-api/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
pinecone
pydantic
sentence-transformers
python-dotenv
requests
